Find the Default value for n in Context

Checks for various information surrounding the call to this function to figure out what value for n should be used

default_n(...)

blueprint_n()

tibble_n()

dplyr_n()

args_n(...)

Arguments

...
parameters to check the length of

Functions

default_n: Checks for the default value of n in a context aware manner.

blueprint_n: If we are inside of a blueprint, then use the n that is passed to the blueprinting function

tibble_n: If we are inside of a call to tibble(), it returns the number of rows in the tibble, otherwise NULL

dplyr_n: If we are in dplyr verb function, it returns the number of rows in the current data, otherwise NULL

args_n: Checks the lengths of the non-NULL arguments passed to it. Errors if they are inconsistent.

Examples

default_n()

#> [1] 10

tibble::tibble(id = 1:3, n = default_n())

#> # A tibble: 3 x 2
#>      id     n
#>   <int> <dbl>
#> 1     1     3
#> 2     2     3
#> 3     3     3

df <- tibble::tibble(id = 1:4)
dplyr::mutate(df, n = default_n())

#> # A tibble: 4 x 2
#>      id     n
#>   <int> <dbl>
#> 1     1     4
#> 2     2     4
#> 3     3     4
#> 4     4     4

default_n(1:5)

#> [1] 5

set_n(10)
default_n()

#> [1] 10

bp <- blueprint(x=r_norm(),n=blueprint_n())
bp(n=10)

#> # A tibble: 10 x 2
#>          x     n
#>      <dbl> <dbl>
#>  1  1.07      10
#>  2  0.0700    10
#>  3 -0.639     10
#>  4 -0.0500    10
#>  5 -0.251     10
#>  6  0.445     10
#>  7  2.76      10
#>  8  0.0465    10
#>  9  0.578     10
#> 10  0.118     10

tibble::tibble(id = 1:3, n = tibble_n())

#> # A tibble: 3 x 2
#>      id     n
#>   <int> <int>
#> 1     1     3
#> 2     2     3
#> 3     3     3

dplyr::mutate(df, n = dplyr_n())

#> # A tibble: 4 x 2
#>      id     n
#>   <int> <int>
#> 1     1     4
#> 2     2     4
#> 3     3     4
#> 4     4     4

args_n(1:3)

#> [1] 3

if (FALSE) {
args_n(1:3, 1:4)
}