Random Seed Defining Functions

Functions related to generating random seeds and utilising them for reproducibility.

gen_seed()

set_seed(seed)

with_seed(seed, expression)

pull_seed(x)

Arguments

seed
The random seed to be used

expression
expression to be evaluated

x
object to extract the seed from

Details

Random values are generated based on the current seed used by the R system. This means by deliberately setting a seed in R, we can make work reproducible.

Functions

gen_seed: Generates a random seed, which can be used in set_seed()

set_seed: Sets the current seed

with_seed: Evaluates the expression after setting the seed. If seed is TRUE, then it first generates a feasible random seed. Results are output with the seed attached (if set), which can be later extracted with pull_seed()

pull_seed: Extracts the seed used to generate the results of with_seed

Examples

my_seed <- gen_seed()
res <- with_seed(my_seed, r_norm(n = 10))
res

#>  [1] -0.3632562 -1.6390605 -0.2930205  1.2217735 -0.4380276  0.5729170
#>  [7]  1.1137839 -0.4031856  0.9239176 -0.5070330
#> attr(,"seed")
#> [1] 377728504

pull_seed(res)

#> [1] 377728504